Try build a update field funciton

+++ field_works.h
@@ -7,8 +7,8 @@
 
 struct arrow_position
 {
-    uint16_t x;
-    uint16_t y;
+    int x;
+    int y;
 };
 
 /*******************Function_Prototypes**********************/
